Move play-ws-1 and play-ws-2 into separate packages
