+++ dd-java-agent/instrumentation/play-ws-1/src/main/java/datadog/trace/instrumentation/playws1/PlayWSClientDecorator.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.playws;
+package datadog.trace.instrumentation.playws1;
 
 import datadog.trace.agent.decorator.HttpClientDecorator;
 import java.net.URI;
