+++ dd-java-agent/instrumentation/play-ws-2/src/main/java/datadog/trace/instrumentation/playws2/HeadersInjectAdapter.java
@@ -1,4 +1,4 @@
-package datadog.trace.instrumentation.playws;
+package datadog.trace.instrumentation.playws2;
 
 import datadog.trace.instrumentation.api.AgentPropagation;
 import play.shaded.ahc.org.asynchttpclient.Request;
